api/requirements.txt
appdirs==1.4.4
attrs==20.3.0
azure-functions==1.6.0
black==20.8b1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
idna==2.10
iniconfig==1.1.1
mongoengine==0.23.0
mypy-extensions==0.4.3
packaging==20.9
pathspec==0.8.1
pluggy==0.13.1
py==1.10.0
pycparser==2.20
PyJWT==2.0.1
pymongo==3.11.3
pyparsing==2.4.7
pytest==6.2.2
python-dotenv==0.15.0
regex==2021.3.17
requests==2.25.1
sentry-sdk==1.0.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
